Add StorageId property to ShopEmployeeDTO and update mapping
The `ShopEmployeeDTO` class in the `HyggyBackend.BLL.DTO.EmployeesDTO`
namespace has been updated to include a new property `StorageId` of
type `long`.

In the `MappingProfiles.cs` file within the `HyggyBackend.BLL.Helper`
namespace, the mapping configuration for `ShopEmployee` to
`ShopEmployeeDTO` has been modified. Specifically, a new mapping has
been added to map the `StorageId` property of `ShopEmployeeDTO` from
the `StorageId` property of the `Shop` associated with the
`ShopEmployee`.